Added client convar querying :o Also added ISourceMod::GetModFolderName() for returning the name of the mod directory by itself
--HG--
extra : convert_revision : svn%3A39bc706e-5318-0410-9160-8a85361fbb7c/trunk%40702

+++ core/vm/sp_vm_basecontext.cpp
@@ -711,7 +711,7 @@ int BaseContext::StringToLocal(cell_t local_addr, size_t bytes, const char *sour
 	len = strlen(source);
 	dest = (char *)(ctx->memory + local_addr);
 
-	if ((size_t)len >= bytes)
+	if (len >= bytes)
 	{
 		len = bytes - 1;
 	}
